+++ src/main/helpers/webstorm-config.js
@@ -9,7 +9,7 @@ const rules = {
 	FileEditorManager              : false,
 	FileTemplateManagerImpl        : false,
 	FindInProjectRecents           : false,
-	'Git.Settings'                 : false,
+	'Git.Settings'                 : true,
 	IdeDocumentHistory             : false,
 	ProjectFrameBounds             : false,
 	ProjectView                    : false,
@@ -47,7 +47,7 @@ export function cleanWorkspace(config) {
 				continue
 			} else if (rule && typeof rule === 'object') {
 				for (const key in rule) {
-					if (Object.prototype.hasOwnProperty.call(rule, key)) {
+					if (Object.prototype.hasOwnProperty.call(rule, key) && rule[key] === false) {
 						delete component[key]
 					}
 				}