+++ src/main/helpers/xml.js
@@ -2,6 +2,7 @@ import path from 'path'
 import fs from 'fs'
 import fse from 'fs-extra'
 import xmlConverter from 'xml-js'
+import format from 'xml-formatter'
 
 export function loadFile(filePath) {
 	return new Promise((resolve, reject) => fs.readFile(filePath, (err, data) => {
@@ -49,6 +50,30 @@ async function loadXml(filePath) {
 	return parseXml(xml)
 }
 
+function traversalObject(object, handler, parentKey) {
+	if (object == null || typeof object !== 'object') {
+		return object
+	}
+
+	for (const key in object) {
+		if (Object.prototype.hasOwnProperty.call(object, key)) {
+			object[key] = handler(key, object[key], parentKey)
+			traversalObject(object[key], handler, key)
+		}
+	}
+
+	return object
+}
+
+function escapeSpecialCharsInAttributes(jsObject) {
+	traversalObject(jsObject, (key, value, parentKey) => {
+		if (parentKey === '_attributes' && typeof value === 'string') {
+			return value.replace(/&/g, '&amp;')
+		}
+		return value
+	})
+}
+
 function saveXml(filePath, jsObject) {
 	const xml = buildXml(jsObject)
 	return saveFileText(filePath, xml)
@@ -59,7 +84,11 @@ function parseXml(xmlText) {
 }
 
 function buildXml(jsObject) {
-	return xmlConverter.js2xml(jsObject, {compact: true})
+	escapeSpecialCharsInAttributes(jsObject)
+	return format(xmlConverter.js2xml(jsObject, {compact: true}), {
+		collapseContent: true,
+		indentation    : '\t',
+	})
 }
 
 export default {